removes all console.logs from Header.js
If it's ok with you, Lars I'd like to progressively
remove stowaway `console.log` statements as part
of my reviews. They make it cumbersome for other
developers to debug their own code.

Instead of leaving a comment I'll just remove them
and if you think they should be there feel free to
`git revert` them (do be so kind to leave a reason
why in the commit log so I know when/when not to
do so in the future).